# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: PortalDots API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080/v1
paths:
...
  /participation-types/{typeID}/registration-form:
    get:
      operationId: getParticipationTypeRegistrationForm
      summary: Return the registration form template for a participation type
      parameters:
        - in: path
          name: typeID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registration form template
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Unauthenticated
        "404":
          description: Participation type not found
  /circles/current:
    put:
      operationId: putCurrentCircle
      summary: Select the current working circle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [circleId]
              properties:
                circleId:
                  type: string
      responses:
        "204":
          description: Current circle updated
        "401":
          description: Unauthenticated
        "404":
          description: Circle not found
        "422":
          description: Validation error
    delete:
      operationId: deleteCurrentCircle
      summary: Delete the current circle for the authenticated user
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthenticated
        "404":
          description: Current circle was not found
        "403":
          description: The current user cannot delete the selected circle
  /circles/current/detail:
    get:
      operationId: getCurrentCircleDetail
      summary: Return the current circle detail
      responses:
        "200":
          description: Current circle detail
          content:
            application/json:
              schema:
                type: object
                required:
                  [
                    id,
                    name,
                    nameYomi,
                    groupName,
                    groupNameYomi,
                    participationTypeId,
                    participationTypeName,
                    notes,
                    invitationToken,
                    submittedAt,
                    status,
                    formCloseAt,
                    places,
                  ]
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  nameYomi:
                    type: string
                  groupName:
                    type: string
                  groupNameYomi:
                    type: string
                  participationTypeId:
                    type: string
                  participationTypeName:
                    type: string
                  notes:
                    type: string
                  invitationToken:
                    type: string
                  submittedAt:
                    anyOf:
                      - type: "null"
                      - type: string
                        format: date-time
                  status:
                    type: string
                    enum: [pending, approved, rejected]
                  formCloseAt:
                    type: string
                    format: date-time
                  places:
                    type: array
                    items:
                      type: string
        "401":
          description: Unauthenticated
        "404":
          description: Current circle was not found
    put:
      operationId: putCurrentCircleDetail
      summary: Update the current circle detail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, nameYomi, groupName, groupNameYomi]
              properties:
                name:
                  type: string
                nameYomi:
                  type: string
                groupName:
                  type: string
                groupNameYomi:
                  type: string
                notes:
                  type: string
                details:
                  type: object
                  additionalProperties:
                    anyOf:
                      - type: string
                      - type: array
                        items:
                          type: string
      responses:
        "200":
          description: Updated current circle
          content:
            application/json:
              schema:
                type: object
                required:
                  [
                    id,
                    name,
                    nameYomi,
                    groupName,
                    groupNameYomi,
                    participationTypeId,
                    participationTypeName,
                    notes,
                    invitationToken,
                    submittedAt,
                  ]
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  nameYomi:
                    type: string
                  groupName:
                    type: string
                  groupNameYomi:
                    type: string
                  participationTypeId:
                    type: string
                  participationTypeName:
                    type: string
                  notes:
                    type: string
                  invitationToken:
                    type: string
                  submittedAt:
                    anyOf:
                      - type: "null"
                      - type: string
                        format: date-time
        "401":
          description: Unauthenticated
        "403":
          description: The current user cannot update the selected circle
        "404":
          description: Current circle was not found
        "422":
          description: Validation error
...